Fix validator's exit code.
Before this commit, the validator ends with a status code at 1 if at
least one validation test failed.

But when different from 0, the CI always fails without publishing the
report in the Wiki. So, it becomes impossible to see the most interesting
part of this validation workflow, especially when it fails:
the validation result.