[IVOA] Remove test with column prefixed with digit
In the DAL Running Meeting #15, it has been decided to remove this
test. Depending on the tokenisation rules of the DBMS, `2weird`
could be interpreted as `2 AS weird` (which would be correct) or
merely rejected by the parser as an error (because a column name
can not start with a digit).